packages/backend/python/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.2
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.7.0
Faker==30.8.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
